# presentation/http/dto/server.gen.go
// Package dto provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package dto

import (
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
)

// ActiveSessionsResponse defines model for ActiveSessionsResponse.
type ActiveSessionsResponse struct {
	Sessions []SessionInfo `json:"sessions"`
}

// ChangeCommandRequest defines model for ChangeCommandRequest.
type ChangeCommandRequest struct {
	// NewWorkName New work name (can be empty)
	NewWorkName string `json:"new_work_name"`

	// UserName User name from Twitch/YouTube
	UserName string `json:"user_name"`
}

// ChangeCommandResponse defines model for ChangeCommandResponse.
type ChangeCommandResponse struct {
	// SessionId Session ID
	SessionId int64 `json:"session_id"`

	// UserId User ID
	UserId int64 `json:"user_id"`

	// WorkName New work name
	WorkName string `json:"work_name"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Error message
	Error string `json:"error"`
}

// JoinCommandRequest defines model for JoinCommandRequest.
type JoinCommandRequest struct {
	// UserName User name from Twitch/YouTube
	UserName string `json:"user_name"`

	// WorkName Optional work description
	WorkName *string `json:"work_name,omitempty"`
}

// JoinCommandResponse defines model for JoinCommandResponse.
type JoinCommandResponse struct {
	// PlannedEnd Planned end time
	PlannedEnd time.Time `json:"planned_end"`

	// SessionId Created session ID
	SessionId int64 `json:"session_id"`

	// StartTime Session start time
	StartTime time.Time `json:"start_time"`

	// UserId User ID
	UserId int64 `json:"user_id"`

	// WorkName Work name (may be empty)
	WorkName *string `json:"work_name,omitempty"`
}

// MoreCommandRequest defines model for MoreCommandRequest.
type MoreCommandRequest struct {
	// Minutes Extension duration in minutes (1-360)
	Minutes int `json:"minutes"`

	// UserName User name from Twitch/YouTube
	UserName string `json:"user_name"`
}

// MoreCommandResponse defines model for MoreCommandResponse.
type MoreCommandResponse struct {
	// Minutes Extended duration in minutes
	Minutes int `json:"minutes"`

	// PlannedEnd New planned end time after extension
	PlannedEnd time.Time `json:"planned_end"`

	// SessionId Extended session ID
	SessionId int64 `json:"session_id"`

	// UserId User ID
	UserId int64 `json:"user_id"`
}

// OutCommandRequest defines model for OutCommandRequest.
type OutCommandRequest struct {
	// UserName User name from Twitch/YouTube
	UserName string `json:"user_name"`
}

// OutCommandResponse defines model for OutCommandResponse.
type OutCommandResponse struct {
	// ActualEnd Session actual end time
	ActualEnd time.Time `json:"actual_end"`

	// SessionId Ended session ID
	SessionId int64 `json:"session_id"`

	// UserId User ID
	UserId int64 `json:"user_id"`
}

// SessionInfo defines model for SessionInfo.
type SessionInfo struct {
	// IconId Icon ID (optional)
	IconId *int64 `json:"icon_id"`

	// PlannedEnd Planned end time
	PlannedEnd time.Time `json:"planned_end"`

	// SessionId Session ID
	SessionId int64 `json:"session_id"`

	// StartTime Session start time
	StartTime time.Time `json:"start_time"`

	// Tier User tier (1, 2, or 3)
	Tier int `json:"tier"`

	// UserId User ID
	UserId int64 `json:"user_id"`

	// UserName User name
	UserName string `json:"user_name"`

	// WorkName Work description
	WorkName string `json:"work_name"`
}

// ChangeCommandJSONRequestBody defines body for ChangeCommand for application/json ContentType.
type ChangeCommandJSONRequestBody = ChangeCommandRequest

// JoinCommandJSONRequestBody defines body for JoinCommand for application/json ContentType.
type JoinCommandJSONRequestBody = JoinCommandRequest

// MoreCommandJSONRequestBody defines body for MoreCommand for application/json ContentType.
type MoreCommandJSONRequestBody = MoreCommandRequest

// OutCommandJSONRequestBody defines body for OutCommand for application/json ContentType.
type OutCommandJSONRequestBody = OutCommandRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Change command (/change)
	// (POST /api/commands/change)
	ChangeCommand(w http.ResponseWriter, r *http.Request)
	// Join command (/in)
	// (POST /api/commands/join)
	JoinCommand(w http.ResponseWriter, r *http.Request)
	// More command (/more)
	// (POST /api/commands/more)
	MoreCommand(w http.ResponseWriter, r *http.Request)
	// Out command (/out)
	// (POST /api/commands/out)
	OutCommand(w http.ResponseWriter, r *http.Request)
	// Get all active sessions
	// (GET /api/sessions/active)
	GetActiveSessions(w http.ResponseWriter, r *http.Request)
	// Health check endpoint
	// (GET /health)
	HealthCheck(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Change command (/change)
// (POST /api/commands/change)
func (_ Unimplemented) ChangeCommand(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Join command (/in)
// (POST /api/commands/join)
func (_ Unimplemented) JoinCommand(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// More command (/more)
// (POST /api/commands/more)
func (_ Unimplemented) MoreCommand(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Out command (/out)
// (POST /api/commands/out)
func (_ Unimplemented) OutCommand(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get all active sessions
// (GET /api/sessions/active)
func (_ Unimplemented) GetActiveSessions(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Health check endpoint
// (GET /health)
func (_ Unimplemented) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ChangeCommand operation middleware
func (siw *ServerInterfaceWrapper) ChangeCommand(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ChangeCommand(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// JoinCommand operation middleware
func (siw *ServerInterfaceWrapper) JoinCommand(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.JoinCommand(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// MoreCommand operation middleware
func (siw *ServerInterfaceWrapper) MoreCommand(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.MoreCommand(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OutCommand operation middleware
func (siw *ServerInterfaceWrapper) OutCommand(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OutCommand(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetActiveSessions operation middleware
func (siw *ServerInterfaceWrapper) GetActiveSessions(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetActiveSessions(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// HealthCheck operation middleware
func (siw *ServerInterfaceWrapper) HealthCheck(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.HealthCheck(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/commands/change", wrapper.ChangeCommand)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/commands/join", wrapper.JoinCommand)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/commands/more", wrapper.MoreCommand)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/commands/out", wrapper.OutCommand)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/sessions/active", wrapper.GetActiveSessions)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/health", wrapper.HealthCheck)
	})

	return r
}