# presentation/http/handler/command_handler.go
package handler

import (
	"encoding/json"
	"net/http"

	"github.com/yamada-ai/workspace-backend/domain"
	"github.com/yamada-ai/workspace-backend/presentation/http/dto"
	"github.com/yamada-ai/workspace-backend/usecase/command"
)

// CommandHandler handles command-related HTTP requests
type CommandHandler struct {
	joinUseCase   *command.JoinCommandUseCase
	outUseCase    *command.OutCommandUseCase
	moreUseCase   *command.MoreCommandUseCase
	changeUseCase *command.ChangeCommandUseCase
}

// NewCommandHandler creates a new command handler
func NewCommandHandler(
	joinUseCase *command.JoinCommandUseCase,
	outUseCase *command.OutCommandUseCase,
	moreUseCase *command.MoreCommandUseCase,
	changeUseCase *command.ChangeCommandUseCase,
) *CommandHandler {
	return &CommandHandler{
		joinUseCase:   joinUseCase,
		outUseCase:    outUseCase,
		moreUseCase:   moreUseCase,
		changeUseCase: changeUseCase,
	}
}

// JoinCommand handles POST /api/commands/join
func (h *CommandHandler) JoinCommand(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req dto.JoinCommandRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "Invalid request body: "+err.Error())
		return
	}

	// Validate user_name
	if req.UserName == "" {
		writeError(w, http.StatusBadRequest, "user_name is required")
		return
	}

	// Prepare usecase input
	workName := ""
	if req.WorkName != nil {
		workName = *req.WorkName
	}

	input := command.JoinCommandInput{
		UserName: req.UserName,
		WorkName: workName,
	}

	// Execute usecase
	output, err := h.joinUseCase.Execute(r.Context(), input)
	if err != nil {
		// Handle already in session error
		if err == domain.ErrUserAlreadyInSession {
			writeError(w, http.StatusConflict, "既に作業セッション中です。先に /out で終了してください。")
			return
		}
		writeError(w, http.StatusInternalServerError, "Failed to join: "+err.Error())
		return
	}

	// Convert to response
	resp := dto.JoinCommandResponse{
		SessionId:  output.SessionID,
		UserId:     output.UserID,
		StartTime:  output.StartTime,
		PlannedEnd: output.PlannedEnd,
	}
	if output.WorkName != "" {
		resp.WorkName = &output.WorkName
	}

	writeJSON(w, http.StatusOK, resp)
}

// OutCommand handles POST /api/commands/out
func (h *CommandHandler) OutCommand(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req dto.OutCommandRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "Invalid request body: "+err.Error())
		return
	}

	// Validate user_name
	if req.UserName == "" {
		writeError(w, http.StatusBadRequest, "user_name is required")
		return
	}

	// Prepare usecase input
	input := command.OutCommandInput{
		UserName: req.UserName,
	}

	// Execute usecase
	output, err := h.outUseCase.Execute(r.Context(), input)
	if err != nil {
		// Handle user not found error
		if err == domain.ErrUserNotFound {
			writeError(w, http.StatusNotFound, "ユーザーが見つかりません。")
			return
		}
		// Handle no active session error
		if err == domain.ErrSessionNotFound {
			writeError(w, http.StatusNotFound, "有効なセッションが見つかりません。")
			return
		}
		writeError(w, http.StatusInternalServerError, "Failed to leave: "+err.Error())
		return
	}

	// Convert to response
	resp := dto.OutCommandResponse{
		SessionId: output.SessionID,
		UserId:    output.UserID,
		ActualEnd: output.ActualEnd,
	}

	writeJSON(w, http.StatusOK, resp)
}

// MoreCommand handles POST /api/commands/more
func (h *CommandHandler) MoreCommand(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req dto.MoreCommandRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "Invalid request body: "+err.Error())
		return
	}

	// Validate user_name
	if req.UserName == "" {
		writeError(w, http.StatusBadRequest, "user_name is required")
		return
	}

	// Validate minutes
	if req.Minutes < command.MinExtensionMinutes || req.Minutes > command.MaxExtensionMinutes {
		writeError(w, http.StatusBadRequest, "minutes must be between 1 and 360")
		return
	}

	// Prepare usecase input
	input := command.MoreCommandInput{
		UserName: req.UserName,
		Minutes:  req.Minutes,
	}

	// Execute usecase
	output, err := h.moreUseCase.Execute(r.Context(), input)
	if err != nil {
		// Handle user not found error
		if err == domain.ErrUserNotFound {
			writeError(w, http.StatusNotFound, "ユーザーが見つかりません。")
			return
		}
		// Handle no active session error
		if err == domain.ErrSessionNotFound {
			writeError(w, http.StatusNotFound, "有効なセッションが見つかりません。")
			return
		}
		// Handle invalid extension error
		if err == domain.ErrInvalidExtension {
			writeError(w, http.StatusBadRequest, "無効な延長時間です。")
			return
		}
		writeError(w, http.StatusInternalServerError, "Failed to extend session: "+err.Error())
		return
	}

	// Convert to response
	resp := dto.MoreCommandResponse{
		SessionId:  output.SessionID,
		UserId:     output.UserID,
		Minutes:    output.Minutes,
		PlannedEnd: output.PlannedEnd,
	}

	writeJSON(w, http.StatusOK, resp)
}

// ChangeCommand handles POST /api/commands/change
func (h *CommandHandler) ChangeCommand(w http.ResponseWriter, r *http.Request) {
	// Parse request body
	var req dto.ChangeCommandRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, http.StatusBadRequest, "Invalid request body: "+err.Error())
		return
	}

	// Validate user_name
	if req.UserName == "" {
		writeError(w, http.StatusBadRequest, "user_name is required")
		return
	}

	// Prepare usecase input
	input := command.ChangeCommandInput{
		UserName:    req.UserName,
		NewWorkName: req.NewWorkName,
	}

	// Execute usecase
	output, err := h.changeUseCase.Execute(r.Context(), input)
	if err != nil {
		// Handle user not found error
		if err == domain.ErrUserNotFound {
			writeError(w, http.StatusNotFound, "ユーザーが見つかりません。")
			return
		}
		// Handle no active session error
		if err == domain.ErrSessionNotFound {
			writeError(w, http.StatusNotFound, "有効なセッションが見つかりません。")
			return
		}
		// Handle session already completed error
		if err == domain.ErrSessionAlreadyCompleted {
			writeError(w, http.StatusBadRequest, "既に完了したセッションです。")
			return
		}
		writeError(w, http.StatusInternalServerError, "Failed to change work name: "+err.Error())
		return
	}

	// Convert to response
	resp := dto.ChangeCommandResponse{
		SessionId: output.SessionID,
		UserId:    output.UserID,
		WorkName:  output.WorkName,
	}

	writeJSON(w, http.StatusOK, resp)
}

// HealthCheck handles GET /health
func (h *CommandHandler) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/plain")
	w.WriteHeader(http.StatusOK)
	if _, err := w.Write([]byte("ok")); err != nil {
		// Log error but don't fail the health check since headers are already sent
		http.Error(w, "Failed to write response", http.StatusInternalServerError)
	}
}

// Helper functions
func writeJSON(w http.ResponseWriter, status int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	if err := json.NewEncoder(w).Encode(data); err != nil {
		// Log error - headers already sent, so we can't change response
		http.Error(w, "Failed to encode JSON", http.StatusInternalServerError)
	}
}

func writeError(w http.ResponseWriter, status int, message string) {
	writeJSON(w, status, dto.ErrorResponse{Error: message})
}