# shared/api/openapi.yaml
openapi: 3.0.3
info:
  title: Workspace Backend API
  version: 1.0.0
  description: API for 24H MMO online coworking space

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /health:
    get:
      summary: Health check endpoint
      operationId: healthCheck
      responses:
        '200':
          description: Server is healthy
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /api/sessions/active:
    get:
      summary: Get all active sessions
      operationId: getActiveSessions
      description: Retrieves all currently active sessions with user information
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveSessionsResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/commands/join:
    post:
      summary: Join command (/in)
      operationId: joinCommand
      description: User joins the workspace and starts a work session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinCommandRequest'
      responses:
        '200':
          description: Successfully joined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinCommandResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User already has an active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/commands/out:
    post:
      summary: Out command (/out)
      operationId: outCommand
      description: User leaves the workspace and ends their work session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OutCommandRequest'
      responses:
        '200':
          description: Successfully left
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OutCommandResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or no active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/commands/more:
    post:
      summary: More command (/more)
      operationId: moreCommand
      description: User extends their current work session by specified minutes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoreCommandRequest'
      responses:
        '200':
          description: Successfully extended session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoreCommandResponse'
        '400':
          description: Bad request (invalid minutes or no active session)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or no active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/commands/change:
    post:
      summary: Change command (/change)
      operationId: changeCommand
      description: User changes their current work session's work name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeCommandRequest'
      responses:
        '200':
          description: Successfully changed work name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeCommandResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or no active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    JoinCommandRequest:
      type: object
      required:
        - user_name
      properties:
        user_name:
          type: string
          description: User name from Twitch/YouTube
          minLength: 1
          maxLength: 100
          example: yamada
        work_name:
          type: string
          description: Optional work description
          maxLength: 200
          example: 論文執筆

    JoinCommandResponse:
      type: object
      required:
        - session_id
        - user_id
        - start_time
        - planned_end
      properties:
        session_id:
          type: integer
          format: int64
          description: Created session ID
          example: 123
        user_id:
          type: integer
          format: int64
          description: User ID
          example: 45
        work_name:
          type: string
          description: Work name (may be empty)
          example: 論文執筆
        start_time:
          type: string
          format: date-time
          description: Session start time
          example: 2025-10-09T14:30:00Z
        planned_end:
          type: string
          format: date-time
          description: Planned end time
          example: 2025-10-09T15:30:00Z

    OutCommandRequest:
      type: object
      required:
        - user_name
      properties:
        user_name:
          type: string
          description: User name from Twitch/YouTube
          minLength: 1
          maxLength: 100
          example: yamada

    OutCommandResponse:
      type: object
      required:
        - session_id
        - user_id
        - actual_end
      properties:
        session_id:
          type: integer
          format: int64
          description: Ended session ID
          example: 123
        user_id:
          type: integer
          format: int64
          description: User ID
          example: 45
        actual_end:
          type: string
          format: date-time
          description: Session actual end time
          example: 2025-10-09T15:25:00Z

    MoreCommandRequest:
      type: object
      required:
        - user_name
        - minutes
      properties:
        user_name:
          type: string
          description: User name from Twitch/YouTube
          minLength: 1
          maxLength: 100
          example: yamada
        minutes:
          type: integer
          description: Extension duration in minutes (1-360)
          minimum: 1
          maximum: 360
          example: 30

    MoreCommandResponse:
      type: object
      required:
        - session_id
        - user_id
        - minutes
        - planned_end
      properties:
        session_id:
          type: integer
          format: int64
          description: Extended session ID
          example: 123
        user_id:
          type: integer
          format: int64
          description: User ID
          example: 45
        minutes:
          type: integer
          description: Extended duration in minutes
          example: 30
        planned_end:
          type: string
          format: date-time
          description: New planned end time after extension
          example: 2025-10-09T16:00:00Z

    ChangeCommandRequest:
      type: object
      required:
        - user_name
        - new_work_name
      properties:
        user_name:
          type: string
          description: User name from Twitch/YouTube
          minLength: 1
          maxLength: 100
          example: yamada
        new_work_name:
          type: string
          description: New work name (can be empty)
          maxLength: 200
          example: 資格勉強

    ChangeCommandResponse:
      type: object
      required:
        - session_id
        - user_id
        - work_name
      properties:
        session_id:
          type: integer
          format: int64
          description: Session ID
          example: 123
        user_id:
          type: integer
          format: int64
          description: User ID
          example: 45
        work_name:
          type: string
          description: New work name
          example: 資格勉強

    SessionInfo:
      type: object
      required:
        - session_id
        - user_id
        - user_name
        - work_name
        - tier
        - start_time
        - planned_end
      properties:
        session_id:
          type: integer
          format: int64
          description: Session ID
          example: 123
        user_id:
          type: integer
          format: int64
          description: User ID
          example: 45
        user_name:
          type: string
          description: User name
          example: yamada
        work_name:
          type: string
          description: Work description
          example: 論文執筆
        tier:
          type: integer
          description: User tier (1, 2, or 3)
          example: 1
        icon_id:
          type: integer
          format: int64
          nullable: true
          description: Icon ID (optional)
          example: 5
        start_time:
          type: string
          format: date-time
          description: Session start time
          example: 2025-10-09T14:30:00Z
        planned_end:
          type: string
          format: date-time
          description: Planned end time
          example: 2025-10-09T15:30:00Z

    ActiveSessionsResponse:
      type: object
      required:
        - sessions
      properties:
        sessions:
          type: array
          items:
            $ref: '#/components/schemas/SessionInfo'

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Invalid tier value